break methods out into files to enable decorating both app and collection
