+++ index.js
@@ -10,6 +10,7 @@
 // require('time-require');
 
 var Base = require('base-methods');
+var decorate = require('./lib/decorate/');
 var helpers = require('./lib/helpers/');
 var utils = require('./lib/utils/');
 var lib = require('./lib/');
@@ -67,51 +68,24 @@ Base.extend(Templates);
 Templates.prototype.defaultConfig = function () {
   // used in plugins to verify the app instance
   this.isApp = true;
-
-  // decorate `option` method onto instance
-  utils.option(this);
+  decorate.init(this);
 
   for (var key in this.options.mixins) {
     this.mixin(key, this.options.mixins[key]);
   }
 
-  this.define('_', {});
-  this.engines = {};
-  lib.helpers(this);
-  this._.engines = new utils.Engines(this.engines);
   this.engine('default', utils.engine);
-
-  this.define('errors', {
-    compile: {
-      callback: 'is sync and does not take a callback function',
-      engine: 'cannot find an engine for: ',
-      method: 'expects engines to have a compile method',
-    },
-    render: {
-      callback: 'is async and expects a callback function',
-      engine: 'cannot find an engine for: ',
-      method: 'expects engines to have a render method',
-    },
-    layouts: {
-      registered: 'no layouts are registered, but one is defined: '
-    }
-  });
-
-  this.cache = {};
-  this.cache.data = {};
-  this.cache.context = {};
-  this.items = {};
-  this.views = {};
-
-  this.viewTypes = {
-    layout: [],
-    renderable: [],
-    partial: []
-  };
-
   this.inflections = {};
   this.listen();
 };
+
+decorate.config(Templates.prototype);
+decorate.routes(Templates.prototype);
+decorate.engine(Templates.prototype);
+decorate.layout(Templates.prototype);
+decorate.render(Templates.prototype);
+decorate.lookup(Templates.prototype);
+decorate.errors(Templates.prototype, 'Templates');
 
 /**
  * Initialize defaults. Exposes constructors on
@@ -173,43 +147,6 @@ Templates.prototype.use = function (fn) {
     this.plugins.push(plugin);
   }
   this.emit('use');
-  return this;
-};
-
-/**
- * Set, get and load data to be passed to templates as
- * context at render-time.
- *
- * ```js
- * app.data('a', 'b');
- * app.data({c: 'd'});
- * console.log(app.cache.data);
- * //=> {a: 'b', c: 'd'}
- * ```
- *
- * @name .data
- * @param {String|Object} `key` Pass a key-value pair or an object to set.
- * @param {any} `val` Any value when a key-value pair is passed. This can also be options if a glob pattern is passed as the first value.
- * @return {Object} Returns an instance of `Templates` for chaining.
- * @api public
- */
-
-Templates.prototype.data = function (key, val) {
-  if (utils.isObject(key)) {
-    this.visit('data', key);
-    return this;
-  }
-
-  var isGlob = typeof val === 'undefined' || utils.hasGlob(key);
-  if (utils.isValidGlob(key) && isGlob) {
-    var opts = utils.extend({}, this.options, val);
-    var data = utils.requireData(key, opts);
-    if (data) this.visit('data', data);
-    return this;
-  }
-
-  key = 'cache.data.' + key;
-  this.set(key, val);
   return this;
 };
 
@@ -401,41 +338,7 @@ Templates.prototype.create = function(name, opts) {
  */
 
 Templates.prototype.extendView = function (view, options) {
-  this.viewDefaults(view, options);
-};
-
-/**
- * Default methods and settings that will be decorated onto
- * each view created by a collection.
- */
-
-Templates.prototype.viewDefaults = function (view, options) {
-  var app = this;
-
-  // decorate `option` method onto `view`
-  if (typeof view.option !== 'function') {
-    utils.option(view);
-    utils.defaults(view.options, options);
-  }
-
-  // decorate `compile` method onto `view`
-  view.compile = function () {
-    var args = [this].concat([].slice.call(arguments));
-    app.compile.apply(app, args);
-    return this;
-  };
-
-  // decorate `render` method onto `view`
-  view.render = function () {
-    var args = [this].concat([].slice.call(arguments));
-    app.render.apply(app, args);
-    return this;
-  };
-
-  // decorate `context` method onto `view`
-  view.context = function(locals) {
-    return utils.merge({}, this.locals, this.data, locals);
-  };
+  decorate.view.all(this, view, options);
 };
 
 /**
@@ -443,829 +346,7 @@ Templates.prototype.viewDefaults = function (view, options) {
  */
 
 Templates.prototype.extendViews = function (views, options) {
-  this.viewsDefaults(views, options);
-};
-
-/**
- * Default methods and settings that will be decorated onto
- * each collection.
- */
-
-Templates.prototype.viewsDefaults = function(views, options) {
-  var app = this;
-
-  if (!views.options.hasOwnProperty('renameKey')) {
-    views.option('renameKey', this.renameKey);
-  }
-
-  views.on('view', function (view) {
-    utils.define(view, 'addView', views.addView.bind(views));
-    app.extendView(view, options);
-
-    app.handleView('onLoad', view);
-    app.emit('view', view);
-  });
-};
-
-/**
- * Set view types for a collection.
- *
- * @param {String} `plural` e.g. `pages`
- * @param {Object} `options`
- */
-
-Templates.prototype.viewType = function(plural, types) {
-  var len = types.length, i = 0;
-  while (len--) {
-    var type = types[i++];
-    this.viewTypes[type] = this.viewTypes[type] || [];
-    if (this.viewTypes[type].indexOf(plural) === -1) {
-      this.viewTypes[type].push(plural);
-    }
-  }
-  return types;
-};
-
-/**
- * Find a view by `name`, optionally passing a `collection` to limit
- * the search. If no collection is passed all `renderable` collections
- * will be searched.
- *
- * ```js
- * var page = app.find('my-page.hbs');
- *
- * // optionally pass a collection name as the second argument
- * var page = app.find('my-page.hbs', 'pages');
- * ```
- * @name .find
- * @param {String} `name` The name/key of the view to find
- * @param {String} `colleciton` Optionally pass a collection name (e.g. pages)
- * @return {Object|undefined} Returns the view if found, or `undefined` if not.
- * @api public
- */
-
-Templates.prototype.find = function (name, collection) {
-  if (typeof name !== 'string') {
-    throw new TypeError('expected name to be a string.');
-  }
-  if (typeof collection === 'string') {
-    return this[collection].getView(name);
-  }
-  var collections = this.viewTypes.renderable;
-  var len = collections.length, i = 0;
-  while (len--) {
-    var plural = collections[i++];
-    var views = this.views[plural];
-    var res;
-    if (res = views[name]) {
-      return res;
-    }
-  }
-};
-
-/**
- * Get view `key` from the specified `collection`.
- *
- * ```js
- * var view = app.getView('pages', 'a/b/c.hbs');
- *
- * // optionally pass a `renameKey` function to modify the lookup
- * var view = app.getView('pages', 'a/b/c.hbs', function(fp) {
- *   return path.basename(fp);
- * });
- * ```
- * @name .getView
- * @param {String} `collection` Collection name, e.g. `pages`
- * @param {String} `key` Template name
- * @param {Function} `fn` Optionally pass a `renameKey` function
- * @return {Object}
- * @api public
- */
-
-Templates.prototype.getView = function(collection, key, fn) {
-  var views = this.getViews(collection);
-  // use custom renameKey function
-  if (typeof fn === 'function') {
-    key = fn(key);
-  }
-  if (views.hasOwnProperty(key)) {
-    return views[key];
-  }
-  // try again with the default renameKey function
-  fn = this.option('renameKey');
-  var name;
-  if (typeof fn === 'function') {
-    name = fn(key);
-  }
-  if (name && name !== key && views.hasOwnProperty(name)) {
-    return views[name];
-  }
-  return null;
-};
-
-/**
- * Get all views from a `collection` using the collection's
- * singular or plural name.
- *
- * ```js
- * var pages = app.getViews('pages');
- * //=> { pages: {'home.hbs': { ... }}
- *
- * var posts = app.getViews('posts');
- * //=> { posts: {'2015-10-10.md': { ... }}
- * ```
- *
- * @name .getViews
- * @param {String} `name` The collection name, e.g. `pages` or `page`
- * @return {Object}
- * @api public
- */
-
-Templates.prototype.getViews = function(name) {
-  var orig = name;
-  if (utils.isObject(name)) return name;
-  if (!this.views.hasOwnProperty(name)) {
-    name = this.inflections[name];
-  }
-  if (!this.views.hasOwnProperty(name)) {
-    throw new Error('getViews cannot find collection: ' + orig);
-  }
-  return this.views[name];
-};
-
-/**
- * Returns the first view from `collection` with a key
- * that matches the given glob pattern.
- *
- * ```js
- * var pages = app.matchView('pages', 'home.*');
- * //=> {'home.hbs': { ... }, ...}
- *
- * var posts = app.matchView('posts', '2010-*');
- * //=> {'2015-10-10.md': { ... }, ...}
- * ```
- *
- * @name .matchView
- * @param {String} `collection` Collection name.
- * @param {String} `pattern` glob pattern
- * @param {Object} `options` options to pass to [micromatch][]
- * @return {Object}
- * @api public
- */
-
-Templates.prototype.matchView = function(collection, pattern, options) {
-  var views = this.getViews(collection);
-  if (views.hasOwnProperty(pattern)) {
-    return views[pattern];
-  }
-  return utils.matchKey(views, pattern, options);
-};
-
-/**
- * Returns any views from the specified collection with keys
- * that match the given glob pattern.
- *
- * ```js
- * var pages = app.matchViews('pages', 'home.*');
- * //=> {'home.hbs': { ... }, ...}
- *
- * var posts = app.matchViews('posts', '2010-*');
- * //=> {'2015-10-10.md': { ... }, ...}
- * ```
- *
- * @name .matchViews
- * @param {String} `collection` Collection name.
- * @param {String} `pattern` glob pattern
- * @param {Object} `options` options to pass to [micromatch]
- * @return {Object}
- * @api public
- */
-
-Templates.prototype.matchViews = function(collection, pattern, options) {
-  var views = this.getViews(collection);
-  return utils.matchKeys(views, pattern, options);
-};
-
-/**
- * Add `Router` and `Route` to the prototype
- */
-
-Templates.prototype.Router = utils.router.Router;
-Templates.prototype.Route = utils.router.Route;
-
-/**
- * Lazily initalize `router`, to allow options to
- * be passed in after init.
- */
-
-Templates.prototype.lazyRouter = function(methods) {
-  if (typeof this.router === 'undefined') {
-    this.define('router', new this.Router({
-      methods: utils.methods
-    }));
-  }
-  if (typeof methods !== 'undefined') {
-    this.router.method(methods);
-  }
-};
-
-/**
- * Handle a middleware `method` for `view`.
- *
- * ```js
- * app.handle('customMethod', view, callback);
- * ```
- * @name .handle
- * @param {String} `method` Name of the router method to handle. See [router methods](./docs/router.md)
- * @param {Object} `view` View object
- * @param {Function} `callback` Callback function
- * @return {Object}
- * @api public
- */
-
-Templates.prototype.handle = function (method, view, locals, cb) {
-  if (typeof locals === 'function') {
-    cb = locals;
-    locals = {};
-  }
-
-  this.lazyRouter();
-  if (!view.options.handled) {
-    view.options.handled = [];
-  }
-
-  if (typeof cb !== 'function') {
-    cb = this.handleError(method, view);
-  }
-
-  view.options.method = method;
-  view.options.handled.push(method);
-  if (view.emit) {
-    view.emit('handle', method);
-  }
-
-  this.router.handle(view, this.handleError(method, view, cb));
-  return this;
-};
-
-/**
- * Run the given middleware handler only if the view has not
- * already been handled by the method.
- *
- * @name .handleView
- * @param  {Object} `method`
- * @param  {Object} `view`
- * @param  {Object} `locals`
- */
-
-Templates.prototype.handleView = function (method, view, locals/*, cb*/) {
-  view.options = view.options || {};
-  if (!view.options.handled) {
-    view.options.handled = [];
-  }
-  if (view.options.handled.indexOf(method) < 0) {
-    this.handle.apply(this, arguments);
-    this.emit(method, view, locals);
-  }
-  return this;
-};
-
-/**
- * Handle middleware errors.
- */
-
-Templates.prototype.handleError = function(method, view, cb) {
-  if (typeof cb !== 'function') cb = utils.noop;
-  var app = this;
-  return function (err) {
-    if (err) {
-      if (err._handled) return cb();
-      err.reason = 'Templates#handle' + method + ': ' + view.path;
-      err._handled = true;
-      app.emit('error', err);
-      return cb(err);
-    }
-    cb(null, view);
-  };
-};
-
-/**
- * Create a new Route for the given path. Each route contains
- * a separate middleware stack.
- *
- * See the [route API documentation][route-api] for details on
- * adding handlers and middleware to routes.
- *
- * ```js
- * app.create('posts');
- * app.route(/blog/)
- *   .all(function(view, next) {
- *     // do something with view
- *     next();
- *   });
- *
- * app.post('whatever', {path: 'blog/foo.bar', content: 'bar baz'});
- * ```
- * @name .route
- * @param {String} `path`
- * @return {Object} `Route` for chaining
- * @api public
- */
-
-Templates.prototype.route = function(path) {
-  this.lazyRouter();
-  return this.router.route.apply(this.router, arguments);
-};
-
-/**
- * Special route method that works just like the `router.METHOD()`
- * methods, except that it matches all verbs.
- *
- * ```js
- * app.all(/\.hbs$/, function(view, next) {
- *   // do stuff to view
- *   next();
- * });
- * ```
- * @name .all
- * @param {String} `path`
- * @param {Function} `callback`
- * @return {Object} `this` for chaining
- * @api public
- */
-
-Templates.prototype.all = function(path/*, callback*/) {
-  var route = this.route(path);
-  route.all.apply(route, [].slice.call(arguments, 1));
-  return this;
-};
-
-/**
- * Add callback triggers to route parameters, where
- * `name` is the name of the parameter and `fn` is the
- * callback function.
- *
- * ```js
- * app.param('title', function (view, next, title) {
- *   //=> title === 'foo.js'
- *   next();
- * });
- *
- * app.onLoad('/blog/:title', function (view, next) {
- *   //=> view.path === '/blog/foo.js'
- *   next();
- * });
- * ```
- * @name .param
- * @param {String} `name`
- * @param {Function} `fn`
- * @return {Object} Returns the instance of `Templates` for chaining.
- * @api public
- */
-
-Templates.prototype.param = function(name/*, fn*/) {
-  this.lazyRouter();
-  this.router.param.apply(this.router, arguments);
-  return this;
-};
-
-/**
- * Register a view engine callback `fn` as `ext`.
- *
- * ```js
- * app.engine('hbs', require('engine-handlebars'));
- *
- * // using consolidate.js
- * var engine = require('consolidate');
- * app.engine('jade', engine.jade);
- * app.engine('swig', engine.swig);
- *
- * // get a registered engine
- * var swig = app.engine('swig');
- * ```
- * @name .engine
- * @param {String|Array} `exts` String or array of file extensions.
- * @param {Function|Object} `fn` or `settings`
- * @param {Object} `settings` Optionally pass engine options as the last argument.
- * @api public
- */
-
-Templates.prototype.engine = function(exts, fn, settings) {
-  if (arguments.length === 1 && typeof exts === 'string') {
-    return this.getEngine(exts);
-  }
-  if (!Array.isArray(exts) && typeof exts !== 'string') {
-    throw new TypeError('expected engine ext to be a string or array.');
-  }
-  utils.arrayify(exts).forEach(function (ext) {
-    this.setEngine(ext, fn, settings);
-  }.bind(this));
-  return this;
-};
-
-/**
- * Register an engine for `ext` with the given `settings`
- *
- * @param {String} `ext` The engine to get.
- */
-
-Templates.prototype.setEngine = function(ext, fn, settings) {
-  ext = utils.formatExt(ext);
-  this._.engines.setEngine(ext, fn, settings);
-  return this;
-};
-
-/**
- * Get the engine settings registered for the given `ext`.
- *
- * @param {String} `ext` The engine to get.
- */
-
-Templates.prototype.getEngine = function(ext) {
-  ext = ext ? utils.formatExt(ext) : null;
-  var engine = this._.engines.getEngine(ext);
-  if (engine) {
-    return engine;
-  }
-  ext = this.option('view engine');
-  return this._.engines.getEngine(ext);
-};
-
-/**
- * Apply a layout to the given `view`.
- *
- * @name .applyLayout
- * @param  {Object} `view`
- * @return {Object} Returns a `view` object.
- */
-
-Templates.prototype.applyLayout = function(view) {
-  if (view.options.layoutApplied) {
-    return view;
-  }
-
-  // handle pre-layout middleware
-  this.handle('preLayout', view);
-
-  // get the layout stack
-  var stack = {}, registered = 0;
-  var alias = this.viewTypes.layout;
-  var len = alias.length, i = 0;
-
-  while (len--) {
-    var views = this.views[alias[i++]];
-    for (var key in views) {
-      stack[key] = views[key];
-      registered++;
-    }
-  }
-
-  // get the name of the first layout
-  var self = this;
-  var name = view.layout;
-  var str = view.content;
-
-  // if no layout is defined, move on
-  if (typeof name === 'undefined') {
-    return view;
-  }
-
-  if (registered === 0 || !stack.hasOwnProperty(name)) {
-    throw this.error('layouts', 'registered', name);
-  }
-
-  var opts = {};
-  utils.extend(opts, this.options);
-  utils.extend(opts, view.options);
-  utils.extend(opts, view.context());
-
-  // Handle each layout before it's applied to a view
-  function handleLayout(obj, stats/*, depth*/) {
-    view.currentLayout = obj.layout;
-    view.define('layoutStack', stats.history);
-    self.handle('onLayout', view);
-    delete view.currentLayout;
-  }
-
-  // actually apply the layout
-  var res = utils.layouts(str, name, stack, opts, handleLayout);
-  if (res.result === str) {
-    throw new Error('layout was not applied to: ' + view.path);
-  }
-
-  view.option('layoutApplied', true);
-  view.option('layoutStack', res.history);
-  view.contents = new Buffer(res.result);
-
-  // handle post-layout middleware
-  this.handle('postLayout', view);
-  return view;
-};
-
-/**
- * Compile `content` with the given `locals`.
- *
- * ```js
- * var indexPage = app.page('some-index-page.hbs');
- * var view = app.compile(indexPage);
- * // view.fn => [function]
- *
- * // you can call the compiled function more than once
- * // to render the view with different data
- * view.fn({title: 'Foo'});
- * view.fn({title: 'Bar'});
- * view.fn({title: 'Baz'});
- * ```
- *
- * @name .compile
- * @param  {Object|String} `view` View object.
- * @param  {Object} `locals`
- * @param  {Boolean} `isAsync` Load async helpers
- * @return {Object} View object with `fn` property with the compiled function.
- * @api public
- */
-
-Templates.prototype.compile = function(view, locals, isAsync) {
-  if (typeof locals === 'boolean') {
-    isAsync = locals;
-    locals = {};
-  }
-
-  if (typeof locals === 'function' || typeof isAsync === 'function') {
-    throw this.error('compile', 'callback');
-  }
-
-  locals = utils.extend({settings: {}}, locals);
-
-  // get the engine to use
-  var ext = utils.resolveEngine(view, locals, this.options);
-  var engine = this.getEngine(ext);
-
-  if (typeof engine === 'undefined') {
-    throw this.error('compile', 'engine', ext);
-  }
-
-  if (engine && engine.options) {
-    locals.settings = utils.merge({}, locals.settings, engine.options);
-  }
-
-  var ctx = view.context(locals);
-
-  // apply layout
-  view = this.applyLayout(view);
-
-  // handle `preCompile` middleware
-  this.handleView('preCompile', view, locals);
-
-  // Bind context to helpers before passing to the engine.
-  this.bindHelpers(view, locals, ctx, (locals.async = isAsync));
-
-  // shallow clone the context and locals
-  var settings = utils.extend({}, ctx, locals);
-  utils.extend(settings, this.mergePartials(settings));
-
-  // compile the string
-  var str = view.contents.toString();
-  view.fn = engine.compile(str, settings);
-
-  // handle `postCompile` middleware
-  this.handleView('postCompile', view, locals);
-  return view;
-};
-
-/**
- * Render a view with the given `locals` and `callback`.
- *
- * ```js
- * var blogPost = app.post.getView('2015-09-01-foo-bar');
- * app.render(blogPost, {title: 'Foo'}, function(err, view) {
- *   // `view` is an object with a rendered `content` property
- * });
- * ```
- * @name .render
- * @param  {Object|String} `view` Instance of `View`
- * @param  {Object} `locals` Locals to pass to template engine.
- * @param  {Function} `callback`
- * @api public
- */
-
-Templates.prototype.render = function (view, locals, cb) {
-  if (typeof locals === 'function') {
-    cb = locals;
-    locals = {};
-  }
-
-  if (typeof cb !== 'function') {
-    throw this.error('render', 'callback');
-  }
-
-  // if `view` is a string, see if it's a cached view
-  if (typeof view === 'string') {
-    view = this.find(view);
-  }
-
-  view.locals = utils.merge({}, view.locals, locals);
-  view.locals = utils.merge({}, this.cache.data, view.locals);
-  var opts = this.options;
-
-  // handle `preRender` middleware
-  this.handleView('preRender', view, locals);
-
-  // get the engine
-  var ext = utils.resolveEngine(view, locals, opts);
-  var engine = this.getEngine(ext);
-
-  if (!engine) {
-    return cb(this.error('render', 'engine', ext));
-  }
-
-  var isAsync = typeof cb === 'function';
-
-  // if it's not already compiled, do that first
-  if (typeof view.fn !== 'function') {
-    try {
-      view = this.compile(view, locals, isAsync);
-      return this.render(view, locals, cb);
-    } catch(err) {
-      this.emit('error', err);
-      return cb.call(this, err);
-    }
-  }
-
-  var ctx = view.context(locals);
-  var context = this.context(view, ctx, locals);
-
-  // render the view
-  return engine.render(view.fn, context, function (err, res) {
-    if (err) {
-      if (opts.rethrow !== false) {
-        err = this.rethrow('render', err, view, context);
-      }
-      this.emit('error', err);
-      return cb.call(this, err);
-    }
-
-    view.contents = res;
-    // handle `postRender` middleware
-    this.handle('postRender', view, locals, cb);
-  }.bind(this));
-};
-
-/**
- * Merge "partials" view types. This is necessary for template
- * engines have no support for partials or only support one
- * type of partials.
- *
- * @name .mergePartials
- * @param {Object} `options` Optionally pass an array of viewTypes to include on `options.viewTypes`
- * @return {Object} Merged partials
- * @api public
- */
-
-Templates.prototype.mergePartials = function (options) {
-  var opts = utils.merge({}, this.options, options);
-  var names = opts.mergeTypes || this.viewTypes.partial;
-  var partials = {};
-  var self = this;
-
-  names.forEach(function (name) {
-    var collection = self.views[name];
-    for (var key in collection) {
-      var view = collection[key];
-
-      // handle `onMerge` middleware
-      self.handleView('onMerge', view);
-
-      if (view.options.nomerge) continue;
-      if (opts.mergePartials !== false) {
-        name = 'partials';
-      }
-
-      // convert the partial to:
-      //=> {'foo.hbs': 'some content...'};
-      partials[name] = partials[name] || {};
-      partials[name][key] = view.content;
-    }
-  });
-  return partials;
-};
-
-/**
- * Build the context for the given `view` and `locals`.
- *
- * @name .context
- * @param  {Object} `view` Templates object
- * @param  {Object} `locals`
- * @return {Object} The object to be passed to engines/views as context.
- */
-
-Templates.prototype.context = function (view, ctx, locals) {
-  var obj = {};
-  utils.extend(obj, ctx);
-  utils.extend(obj, this.cache.data);
-  utils.extend(obj, view.locals);
-  utils.extend(obj, view.data);
-  utils.extend(obj, locals);
-  obj.view = view;
-  return obj;
-};
-
-/**
- * Bind context to helpers.
- */
-
-Templates.prototype.bindHelpers = function (view, locals, context, isAsync) {
-  var helpers = {};
-  utils.extend(helpers, this.options.helpers);
-  utils.extend(helpers, this._.helpers.sync);
-
-  if (isAsync) utils.extend(helpers, this._.helpers.async);
-  utils.extend(helpers, locals.helpers);
-
-  var options = {};
-
-  // support helper options: `app.option('helper.foo', 'bar')`
-  if (this.options.hasOwnProperty('helper')) {
-    var opts = this.options.helper;
-    for (var key in opts) {
-      if (opts.hasOwnProperty(key) && helpers.hasOwnProperty(key)) {
-        options[key] = opts[key];
-      }
-    }
-  }
-
-  // build the context to expose as `this` in helpers
-  var thisArg = {};
-  thisArg.options = utils.extend({}, this.options, options, locals);
-  thisArg.context = context || {};
-  thisArg.context.view = view;
-  thisArg.app = this;
-
-  // bind template helpers to the instance
-  locals.helpers = utils.bindAll(helpers, thisArg);
-};
-
-/**
- * Add a router handler.
- *
- * @param  {String} `method` Method name.
- */
-
-Templates.prototype.handler = function (methods) {
-  this.handlers(methods);
-};
-
-/**
- * Add default Router handlers to Templates.
- */
-
-Templates.prototype.handlers = function (methods) {
-  this.lazyRouter(methods);
-  methods.forEach(function (method) {
-    this.define(method, function(path) {
-      var route = this.route(path);
-      var args = [].slice.call(arguments, 1);
-      route[method].apply(route, args);
-      return this;
-    }.bind(this));
-  }.bind(this));
-};
-
-/**
- * Format an error
- *
- * TODO:
- *  - add engine info to error
- */
-
-Templates.prototype.error = function(method, id, msg) {
-  var ctx = this.errors[method][id];
-  var reason = 'Templates#' + method + ' ' + ctx + (msg || '');
-  var err = new Error(reason);
-  err.reason = reason;
-  err.id = id;
-  err.msg = msg;
-  return err;
-};
-
-/**
- * Rethrow an error in the given context to
- * get better error messages.
- */
-
-Templates.prototype.rethrow = function(method, err, view, context) {
-  try {
-    utils.rethrow(view.contents.toString(), {
-      data: context,
-      fp: view.path
-    });
-  } catch (msg) {
-    err.method = method;
-    err.reason = msg;
-    err.id = 'rethrow';
-    err._called = true;
-    return err;
-  }
+  decorate.views(this, views, options);
 };
 
 /**
@@ -1275,16 +356,6 @@ Templates.prototype.rethrow = function(method, err, view, context) {
 Templates.prototype.mixin = function(key, value) {
   Templates.prototype[key] = value;
 };
-
-// Add router methods to Templates
-utils.methods.forEach(function (method) {
-  Templates.prototype[method] = function(path) {
-    var route = this.route(path);
-    var args = [].slice.call(arguments, 1);
-    route[method].apply(route, args);
-    return this;
-  };
-});
 
 /**
  * Expose `Templates`
