+++ lib/collection.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var Base = require('base-methods');
+var decorate = require('./decorate');
 var utils = require('./utils');
 
 /**
@@ -20,7 +21,6 @@ function Collection(options) {
 
   // decorate the instance
   utils.renameKey(this);
-  utils.option(this);
 
   // add constructors
   this.define('List', options.List || require('./list'));
@@ -47,6 +47,8 @@ function Collection(options) {
 }
 
 Base.extend(Collection);
+
+decorate.config(Collection.prototype);
 
 /**
  * Run a plugin on the collection instance. Plugins
