+++ lib/group.js
@@ -58,7 +58,7 @@ Group.prototype.use = function(fn) {
  * @return {[type]}
  */
 
-Group.prototype.get = function(key) {
+Group.prototype.get = function() {
   var res = Base.prototype.get.apply(this, arguments);
   if (Array.isArray(res)) {
     var List = this.List;
