+++ lib/item.js
@@ -1,9 +1,10 @@
 'use strict';
 
+var Vinyl = require('vinyl');
+var Stream = require('stream');
 var Base = require('base-methods');
 var cloneStats = require('clone-stats');
-var Stream = require('stream');
-var Vinyl = require('vinyl');
+var decorate = require('./decorate');
 var utils = require('./utils/');
 
 /**
@@ -28,7 +29,6 @@ function Item(item) {
   this.options = item.options || {};
   this.locals = item.locals || {};
   this.data = item.data || {};
-  utils.option(item);
 
   this.define('_name', 'Item');
   this.define('_contents', null);
@@ -63,6 +63,7 @@ function Item(item) {
 
 Base.extend(Item);
 Base.inherit(Item, Vinyl);
+decorate.config(Item.prototype);
 
 /**
  * Run a plugin on the `item` instance.
