+++ lib/list.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var Base = require('base-methods');
+var decorate = require('./decorate');
 var utils = require('./utils');
 var Group = require('./group');
 
@@ -10,7 +11,6 @@ function List(options) {
 
   // decorate the instance
   utils.renameKey(this);
-  utils.option(this);
 
   // add constructors to the instance
   this.define('Item', options.Item || require('./view'));
@@ -41,6 +41,7 @@ function List(options) {
  */
 
 Base.extend(List);
+decorate.config(List.prototype);
 
 /**
  * Run a plugin on the list instance. Plugins
