+++ lib/utils/proto.js
@@ -23,35 +23,8 @@ utils.isItem = utils.isView = function isView(val) {
 };
 
 /**
- * Resolve the name of the engine to use, or the file
- * extension to use for the engine.
- *
- * If `options.resolveEngine` is a function, it will be
- * used to resolve the engine.
- *
- * @param {Object} `view`
- * @param {Object} `locals`
- * @param {Object} `opts`
- * @return {String}
- */
-
-utils.resolveEngine = function resolveEngine(view, locals, opts) {
-  var engine = locals.engine
-    || view.options.engine
-    || view.engine
-    || view.data.engine
-    || (view.data.ext = path.extname(view.path));
-
-  var fn = opts.resolveEngine
-    || locals.resolveEngine
-    || common.identity;
-
-  return fn(engine);
-};
-
-/**
  * Sync the _content and _contents properties on a view to ensure
- * both are set when setting one.
+ * both are set when setting either.
  *
  * @param {Object} `view` instance of a `View`
  * @param {String|Buffer|Stream|null} `contents` contents to set on both properties
@@ -123,29 +96,6 @@ utils.itemFactory = function (app, method, CtorName) {
 };
 
 /**
- * Set or get an option value. This is a factory for
- * adding an `option` method to a class
- */
-
-utils.option = function option(app) {
-  lazy.define(app, 'option', function(key, value) {
-    if (typeof key === 'string') {
-      if (arguments.length === 1) {
-        return this.get('options.' + key);
-      }
-      this.set('options.' + key, value);
-      this.emit('option', key, value);
-      return this;
-    }
-    if (typeof key !== 'object') {
-      throw new TypeError('expected a string or object.');
-    }
-    this.visit('option', key);
-    return this;
-  }.bind(app));
-};
-
-/**
  * Get locals from helper arguments.
  *
  * @param  {Object} `locals`
@@ -171,7 +121,34 @@ utils.getLocals = function getLocals(locals, options) {
 };
 
 /**
- * Resolve the layout to use for `view`. If `options.layoutFn` is
+ * Resolve the name of the engine to use, or the file
+ * extension to use for the engine.
+ *
+ * If `options.resolveEngine` is a function, it will be
+ * used to resolve the engine.
+ *
+ * @param {Object} `view`
+ * @param {Object} `locals`
+ * @param {Object} `opts`
+ * @return {String}
+ */
+
+utils.resolveEngine = function resolveEngine(view, locals, opts) {
+  var engine = locals.engine
+    || opts.engine
+    || view.options.engine
+    || view.engine
+    || view.data.engine
+    || (view.data.ext = path.extname(view.path));
+
+  var fn = opts.resolveEngine
+    || locals.resolveEngine
+    || common.identity;
+  return fn(engine);
+};
+
+/**
+ * Resolve the layout to use for `view`. If `options.resolveLayout` is
  * a function, it will be used instead.
  *
  * @param {Object} `view`
@@ -179,9 +156,9 @@ utils.getLocals = function getLocals(locals, options) {
  * @api public
  */
 
-utils.layoutFn = function layoutFn(view) {
-  if (view.options && typeof view.options.layoutFn === 'function') {
-    return view.options.layoutFn(view);
+utils.resolveLayout = function resolveLayout(view) {
+  if (view.options && typeof view.options.resolveLayout === 'function') {
+    return view.options.resolveLayout(view);
   }
   return view.data.layout || view.locals.layout || view.options.layout;
 };
