+++ lib/view.js
@@ -94,7 +94,7 @@ utils.define(View.prototype, 'layout', {
     this.define('_layout', val);
   },
   get: function () {
-    return this._layout || utils.layoutFn(this);
+    return this._layout || utils.resolveLayout(this);
   }
 });
