+++ lib/views.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var Base = require('base-methods');
+var decorate = require('./decorate');
 var utils = require('./utils');
 
 /**
@@ -19,8 +20,8 @@ function Views(options) {
   options = options || {};
 
   // decorate the instance
+  decorate.init(this);
   utils.renameKey(this);
-  utils.option(this);
 
   // add constructors
   this.define('List', options.List || require('./list'));
@@ -47,6 +48,14 @@ function Views(options) {
 }
 
 Base.extend(Views);
+
+decorate.config(Views.prototype);
+decorate.routes(Views.prototype);
+decorate.engine(Views.prototype);
+decorate.layout(Views.prototype);
+decorate.render(Views.prototype);
+decorate.lookup(Views.prototype);
+decorate.errors(Views.prototype, 'Views');
 
 /**
  * Run a plugin on the collection instance. Plugins
@@ -135,6 +144,8 @@ Views.prototype.addView = function(key, value) {
   while (this.queue.length) {
     this.setView(this.queue.shift());
   }
+
+  this.extendView(view);
   return view;
 };
 
@@ -212,6 +223,10 @@ Views.prototype.addList = function(list) {
 
 Views.prototype.getView = function(key) {
   return this.views[key] || this.views[this.renameKey(key)];
+};
+
+Views.prototype.extendView = function(view) {
+  decorate.view.all(this, view);
 };
 
 /**
